requirements.txt
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.19.0
